+++ liblzma-rs/src/common/stream_flags_decoder.rs
@@ -40,13 +40,13 @@ pub unsafe extern "C" fn lzma_stream_header_decode(
         return LZMA_FORMAT_ERROR;
     }
     let crc: u32 = lzma_crc32(
-        in_0.offset(core::mem::size_of::<[u8; 6]>() as usize as isize),
+        in_0.offset(core::mem::size_of::<[u8; 6]>() as isize),
         LZMA_STREAM_FLAGS_SIZE as size_t,
         0,
     ) as u32;
     if crc
         != read32le(
-            in_0.offset(core::mem::size_of::<[u8; 6]>() as usize as isize)
+            in_0.offset(core::mem::size_of::<[u8; 6]>() as isize)
                 .offset(LZMA_STREAM_FLAGS_SIZE as isize),
         )
     {
@@ -54,7 +54,7 @@ pub unsafe extern "C" fn lzma_stream_header_decode(
     }
     if stream_flags_decode(
         options,
-        in_0.offset(core::mem::size_of::<[u8; 6]>() as usize as isize),
+        in_0.offset(core::mem::size_of::<[u8; 6]>() as isize),
     ) {
         return LZMA_OPTIONS_ERROR;
     }
@@ -67,7 +67,7 @@ pub unsafe extern "C" fn lzma_stream_footer_decode(
     in_0: *const u8,
 ) -> lzma_ret {
     if memcmp(
-        in_0.offset((core::mem::size_of::<u32>() as usize).wrapping_mul(2 as usize) as isize)
+        in_0.offset((core::mem::size_of::<u32>()).wrapping_mul(2) as isize)
             .offset(LZMA_STREAM_FLAGS_SIZE as isize) as *const c_void,
         &raw const lzma_footer_magic as *const u8 as *const c_void,
         core::mem::size_of::<[u8; 2]>(),
@@ -76,7 +76,7 @@ pub unsafe extern "C" fn lzma_stream_footer_decode(
         return LZMA_FORMAT_ERROR;
     }
     let crc: u32 = lzma_crc32(
-        in_0.offset(core::mem::size_of::<u32>() as usize as isize),
+        in_0.offset(core::mem::size_of::<u32>() as isize),
         (core::mem::size_of::<u32>()).wrapping_add(LZMA_STREAM_FLAGS_SIZE as size_t),
         0,
     ) as u32;
@@ -85,12 +85,12 @@ pub unsafe extern "C" fn lzma_stream_footer_decode(
     }
     if stream_flags_decode(
         options,
-        in_0.offset((core::mem::size_of::<u32>() as usize).wrapping_mul(2 as usize) as isize),
+        in_0.offset((core::mem::size_of::<u32>()).wrapping_mul(2) as isize),
     ) {
         return LZMA_OPTIONS_ERROR;
     }
     (*options).backward_size =
-        read32le(in_0.offset(core::mem::size_of::<u32>() as usize as isize)) as lzma_vli;
+        read32le(in_0.offset(core::mem::size_of::<u32>() as isize)) as lzma_vli;
     (*options).backward_size = (*options).backward_size.wrapping_add(1).wrapping_mul(4);
     LZMA_OK
 }
